backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
nltk==3.8.1
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
